+++ lib/apex_charts/options_builder.rb
@@ -240,7 +240,7 @@ module ApexCharts
       @built[:theme] = if theme.is_a? String
                          case theme
                          when 'random'
-                           resolve_theme(Theme::Local.all_palettes.sample)
+                           resolve_theme(Theme.all_palettes.sample)
                          when 'monochrome'
                            {monochrome: {enabled: true}}
                          else
@@ -336,8 +336,8 @@ module ApexCharts
     def resolve_theme(theme)
       if Theme::PALETTES.include? theme
         {palette: theme}
-      elsif Theme::Local.palette_names.include? theme
-        @built[:colors] = Theme::Local.get_colors(theme)
+      elsif Theme.palette_names.include? theme
+        @built[:colors] = Theme.get_colors(theme)
         nil
       end
     end