+++ spec/theme_spec.rb
@@ -2,7 +2,13 @@
 
 require 'spec_helper'
 
-RSpec.shared_examples 'apexcharts_theme' do
+RSpec.describe ApexCharts::Theme do
+  context '::PALETTES' do
+    it 'has 10 palettes' do
+      expect(described_class::PALETTES.size).to eq 10
+    end
+  end
+
   context '.create' do
     it 'creates a new palette' do
       described_class.create(
@@ -63,40 +69,6 @@ RSpec.shared_examples 'apexcharts_theme' do
       expect(described_class.all_palettes).to include 'my_palette'
 
       described_class.destroy 'my_palette'
-    end
-  end
-end
-
-RSpec.describe ApexCharts::Theme do
-  it_behaves_like 'apexcharts_theme'
-
-  context '::PALETTES' do
-    it 'has 10 palettes' do
-      expect(described_class::PALETTES.size).to eq 10
-    end
-  end
-
-  describe described_class::Local do
-    it_behaves_like 'apexcharts_theme'
-
-    after do
-      Thread.current[:_ApexCharts_Palettes_] = {}
-    end
-
-    it 'calls a hash on current thread variable' do
-      expect(Thread.current).to(
-        receive(:[]).with(:_ApexCharts_Palettes_).and_return({})
-      )
-      described_class.palettes
-    end
-
-    it 'works only on current thread' do
-      described_class.create 'local_theme', ['#aabbcc', '#bbccdd', '#ccddee']
-      expect(described_class.all_palettes.size).to eq 11
-      Thread.new {
-        expect(described_class.all_palettes.size).to eq 10
-      }
-      expect(described_class.all_palettes.size).to eq 11
     end
   end
 end