+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Changed
 - Make obsolete previously deprecated local palette creation
-- Update options schema to match ApexCharts.JS version 3.22.0
-- ApexCharts.JS version from 3.21.0 to 3.22.0
+- Update options schema to match ApexCharts.JS version 3.27.3
+- ApexCharts.JS version from 3.21.0 to 3.27.3
 - Add alternative options schema :dry_schema
+- **Box Plot chart**
 
 ## [0.1.10] - 2020-10-04
 ### Added
@@ -174,7 +175,7 @@ All notable changes to this project will be documented in this file.
 - Travis CI
 - **Donut chart**
 - **Pie chart**
-- **RadialBar chart**
+- **Radial Bar chart**
 - Chart examples
 - Logo